Rename logic.py -> ai.py
